val type update

+++ src/DesignCompile/DesignElaboration.cpp
@@ -965,8 +965,8 @@ void DesignElaboration::elaborateInstance_(
         NodeId constExpr = fC->Sibling(varId);
 
         bool validValue;
-        int64_t initVal =
-            m_helper.getValue(validValue, def, fC, constExpr, m_compileDesign,
+        uint64_t initVal =
+            (uint64_t) m_helper.getValue(validValue, def, fC, constExpr, m_compileDesign,
                               nullptr, parent, true, false);
         Value* initValue = m_exprBuilder.getValueFactory().newLValue();
         initValue->set(initVal);
